Ensure client in src is usable
This changes the approach used to test the client by providing a client wrapper which provides the node-specific bits of the environment.
The net result is we don't do the templating changes in the build to produce two versions of the file - so the .js file in git is usable without having to run maven.

Have also added the generation of a compressed version of the file in the build output.

Move test files up a level

Add README.md