Support browser usage in strict mode
Currently, when passing `this` inside the scope setup function in a
browser leads to a `Cannot set property Paho of undefined` error.

To address this we wrap up the initialization logic in another scoped
IIFE that sets up a commonjsGlobal variable with appropriate fallback.

We can then use this variable instead of the undefined `this` when we
attempt to export the Paho library.

Signed-off-by: Matt Cowan <mcowan@atlassian.com>